refactor: modularize extract.py into structured extractor package and update docs
