+++ scripts/extractor/__init__.py
@@ -1,0 +1,1 @@
+from extractor.utils import resolve_input_files, extract_single_file, main
