+++ scripts/extractor/dependencies.py
@@ -1,0 +1,151 @@
+import importlib.util
+import os
+import shutil
+import subprocess
+import sys
+from extractor.config import PYTHON_DEPENDENCIES, HTML_EXTENSIONS
+
+
+def python_module_available(module_name: str) -> bool:
+    return importlib.util.find_spec(module_name) is not None
+
+
+def missing_python_packages(module_names: list[str]) -> list[str]:
+    missing = []
+    for module_name in module_names:
+        if not python_module_available(module_name):
+            missing.append(PYTHON_DEPENDENCIES[module_name])
+    return missing
+
+
+def install_python_packages(packages: list[str]) -> bool:
+    if not packages:
+        return True
+
+    print(f"Installing missing Python package(s): {', '.join(packages)}")
+    try:
+        result = subprocess.run(
+            [sys.executable, "-m", "pip", "install", *packages],
+            text=True,
+            timeout=600,
+        )
+    except Exception as exc:
+        print(f"Package installation failed: {exc}", file=sys.stderr)
+        return False
+
+    importlib.invalidate_caches()
+    return result.returncode == 0
+
+
+def normalize_install_mode(argv: list[str]) -> str:
+    mode = os.environ.get("BOOK_SKILL_INSTALL_MISSING", "ask").lower()
+    if "--no-install-missing" in argv:
+        return "no"
+    if "--install-missing" in argv:
+        idx = argv.index("--install-missing")
+        if idx + 1 < len(argv) and not argv[idx + 1].startswith("--"):
+            mode = argv[idx + 1].lower()
+        else:
+            mode = "yes"
+    if mode in {"1", "true", "y", "yes", "install"}:
+        return "yes"
+    if mode in {"0", "false", "n", "no", "fallback", "skip"}:
+        return "no"
+    return "ask"
+
+
+def offer_dependency_install(
+    *,
+    feature: str,
+    module_names: list[str],
+    fallback: str | None,
+    install_mode: str,
+) -> None:
+    packages = missing_python_packages(module_names)
+    if not packages:
+        return
+
+    message = f"{feature} uses {', '.join(packages)} if installed"
+    if fallback:
+        message += f", otherwise {fallback}"
+    message += "."
+    print(message)
+
+    should_install = False
+    if install_mode == "yes":
+        should_install = True
+    elif install_mode == "ask" and sys.stdin.isatty():
+        answer = input("Missing package(s) detected. Do you want to install? y=install, n=fallback: ").strip().lower()
+        should_install = answer in {"y", "yes", "install"}
+    else:
+        if fallback:
+            print("Non-interactive mode or install disabled; using fallback.")
+        else:
+            print("Non-interactive mode or install disabled; installation skipped.")
+
+    if not should_install:
+        if fallback:
+            print(f"Using fallback: {fallback}.")
+        return
+
+    if install_python_packages(packages):
+        still_missing = missing_python_packages(module_names)
+        if not still_missing:
+            print("Package installation complete.")
+            return
+        print(f"Package installation incomplete; still missing: {', '.join(still_missing)}", file=sys.stderr)
+    else:
+        print("Package installation failed.", file=sys.stderr)
+
+    if fallback:
+        print(f"Using fallback: {fallback}.")
+
+
+def prepare_dependencies(ext: str, extraction_mode: str, install_mode: str) -> None:
+    if ext == ".pdf" and extraction_mode == "technical":
+        offer_dependency_install(
+            feature="Technical PDF extraction",
+            module_names=["docling"],
+            fallback="the PDF text fallback chain",
+            install_mode=install_mode,
+        )
+
+    if ext == ".pdf" and not shutil.which("pdftotext"):
+        offer_dependency_install(
+            feature="PDF text extraction",
+            module_names=["PyPDF2", "pdfminer"],
+            fallback="any installed Python PDF parser; extraction fails if none are available",
+            install_mode=install_mode,
+        )
+
+    if ext == ".epub":
+        offer_dependency_install(
+            feature="EPUB extraction",
+            module_names=["ebooklib", "bs4"],
+            fallback="a stdlib ZIP/HTML parser",
+            install_mode=install_mode,
+        )
+
+    if ext in HTML_EXTENSIONS:
+        offer_dependency_install(
+            feature="HTML extraction",
+            module_names=["bs4"],
+            fallback="a stdlib HTML parser",
+            install_mode=install_mode,
+        )
+
+    if ext == ".docx":
+        offer_dependency_install(
+            feature="DOCX extraction",
+            module_names=["docx"],
+            fallback="a stdlib ZIP/XML parser",
+            install_mode=install_mode,
+        )
+
+    if ext == ".rtf":
+        offer_dependency_install(
+            feature="RTF extraction",
+            module_names=["striprtf"],
+            fallback="a basic regex cleanup fallback",
+            install_mode=install_mode,
+        )
