+++ scripts/extractor/parsers/__init__.py
@@ -1,0 +1,1 @@
+# Parsers package
