+++ scripts/extractor/parsers/docx.py
@@ -1,0 +1,62 @@
+import sys
+import zipfile
+from extractor.config import supported_formats_message
+
+
+def extract_docx_with_python_docx(docx_path: str) -> str | None:
+    try:
+        import docx
+        document = docx.Document(docx_path)
+        parts = [paragraph.text for paragraph in document.paragraphs if paragraph.text]
+        for table in document.tables:
+            for row in table.rows:
+                cells = [cell.text.strip() for cell in row.cells]
+                if any(cells):
+                    parts.append("\t".join(cells))
+        return "\n".join(parts)
+    except ImportError:
+        return None
+    except Exception:
+        return None
+
+
+def extract_docx_with_zipfile(docx_path: str) -> str | None:
+    try:
+        import xml.etree.ElementTree as ET
+
+        with zipfile.ZipFile(docx_path) as zf:
+            xml_bytes = zf.read("word/document.xml")
+        root = ET.fromstring(xml_bytes)
+        namespace = "{http://schemas.openxmlformats.org/wordprocessingml/2006/main}"
+        parts: list[str] = []
+        for paragraph in root.iter(f"{namespace}p"):
+            texts = [node.text for node in paragraph.iter(f"{namespace}t") if node.text]
+            if texts:
+                parts.append("".join(texts))
+        return "\n".join(parts) if parts else None
+    except Exception:
+        return None
+
+
+def extract_docx(docx_path: str) -> tuple[str, str]:
+    print("Trying python-docx...", end=" ", flush=True)
+    text = extract_docx_with_python_docx(docx_path)
+    if text and text.strip():
+        print("OK")
+        return text, "python-docx"
+
+    print("not available")
+    print("Trying stdlib DOCX parser...", end=" ", flush=True)
+    text = extract_docx_with_zipfile(docx_path)
+    if text and text.strip():
+        print("OK")
+        return text, "zipfile-docx"
+
+    print("FAILED")
+    print(
+        "\nERROR: Could not extract text from DOCX.\n"
+        "Install python-docx for best results:\n"
+        "  pip3 install python-docx",
+        file=sys.stderr,
+    )
+    sys.exit(1)
