+++ scripts/extractor/parsers/epub.py
@@ -1,0 +1,66 @@
+import re
+import zipfile
+from extractor.parsers.html import _HTMLTextExtractor
+
+
+def extract_with_ebooklib(epub_path: str) -> str | None:
+    try:
+        import ebooklib
+        from ebooklib import epub
+        from bs4 import BeautifulSoup
+
+        book = epub.read_epub(epub_path)
+        parts = []
+        for item in book.get_items_of_type(ebooklib.ITEM_DOCUMENT):
+            soup = BeautifulSoup(item.get_content(), "html.parser")
+            parts.append(soup.get_text(separator="\n"))
+        return "\n\n".join(parts)
+    except ImportError:
+        return None
+    except Exception:
+        return None
+
+
+def extract_with_zipfile(epub_path: str) -> str | None:
+    """stdlib-only EPUB extractor: unzip → parse HTML files."""
+    try:
+        with zipfile.ZipFile(epub_path) as zf:
+            names = zf.namelist()
+            # Read OPF spine to get reading order, fall back to sorted xhtml files
+            spine_order: list[str] = []
+            opf_files = [n for n in names if n.endswith(".opf")]
+            if opf_files:
+                opf_text = zf.read(opf_files[0]).decode("utf-8", errors="replace")
+                spine_order = re.findall(r'href=["\']([^"\']+\.(?:xhtml|html))["\']', opf_text)
+
+            html_files = spine_order or sorted(
+                n for n in names if n.endswith((".html", ".xhtml"))
+            )
+            if not html_files:
+                return None
+
+            parts = []
+            for name in html_files:
+                try:
+                    raw = zf.read(name).decode("utf-8", errors="replace")
+                    parser = _HTMLTextExtractor()
+                    parser.feed(raw)
+                    parts.append(parser.get_text())
+                except Exception:
+                    continue
+            return "\n\n".join(parts) if parts else None
+    except Exception:
+        return None
+
+
+def count_epub_chapters(epub_path: str) -> int:
+    """Count spine items (approximate chapter count) without dependencies."""
+    try:
+        with zipfile.ZipFile(epub_path) as zf:
+            opf_files = [n for n in zf.namelist() if n.endswith(".opf")]
+            if not opf_files:
+                return 0
+            opf_text = zf.read(opf_files[0]).decode("utf-8", errors="replace")
+            return len(re.findall(r'<itemref\b', opf_text))
+    except Exception:
+        return 0
