+++ scripts/extractor/parsers/html.py
@@ -1,0 +1,52 @@
+import html
+import html.parser
+from extractor.parsers.text import read_text_file
+
+
+class _HTMLTextExtractor(html.parser.HTMLParser):
+    """Minimal HTML → plain text converter using stdlib only."""
+
+    SKIP_TAGS = {"script", "style", "head"}
+
+    def __init__(self):
+        super().__init__()
+        self._parts: list[str] = []
+        self._skip_depth = 0
+        self._current_skip: str | None = None
+
+    def handle_starttag(self, tag, attrs):
+        if tag in self.SKIP_TAGS:
+            self._skip_depth += 1
+        if tag in ("p", "br", "h1", "h2", "h3", "h4", "h5", "h6", "li", "div"):
+            self._parts.append("\n")
+
+    def handle_endtag(self, tag):
+        if tag in self.SKIP_TAGS and self._skip_depth:
+            self._skip_depth -= 1
+
+    def handle_data(self, data):
+        if not self._skip_depth:
+            self._parts.append(data)
+
+    def get_text(self) -> str:
+        return html.unescape("".join(self._parts))
+
+
+def extract_html_content(raw_html: str) -> str:
+    try:
+        from bs4 import BeautifulSoup
+        soup = BeautifulSoup(raw_html, "html.parser")
+        for element in soup(["script", "style", "head"]):
+            element.decompose()
+        return soup.get_text(separator="\n")
+    except ImportError:
+        parser = _HTMLTextExtractor()
+        parser.feed(raw_html)
+        return parser.get_text()
+
+
+def extract_html_file(path: str) -> str | None:
+    raw = read_text_file(path)
+    if raw is None:
+        return None
+    return extract_html_content(raw)
